requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
openai==1.50.0
google-cloud-texttospeech==2.16.0
google-auth==2.29.0
moviepy==1.0.3
pillow==10.4.0
requests==2.32.0
python-multipart==0.0.9
